feat(Detector): Refactoring Detector in AutozoomCtl and Detector classes
Splitting up responsibilities to AutozoomCtl and Detector classes where the former is responsible
for configuring the autozoom (genius framing) feature, wherease the later is responsible for
subscribing to "DETECTIONS" and "FRAMING" events emitted from the camera.

BREAKING CHANGE: The breaking change includes separate classes for autozoom control and detector
data setup. The IDeviceManager now has a `getDetector` and `getAutozoomCtl` method and should be
used distinctively. The `Detector` class does not have the old `enable`,`disable`, `start` and
`stop` commands! They are moved to `AutozoomCtl` instead. The detector class will not work unless
autozoom or genius framing has been properly configured (Documentation and example code snipets are
available in the official SDK docs page).

fix #169
