+++ docs/additional-documentation/rest-api.html
@@ -62,8 +62,8 @@
 </ul>
 <h2 id="detector">Detector</h2>
 <ul>
-<li><a href="rest-api/detector-start.html">Start genius framing and detector</a> : <code>PUT /detector/start</code></li>
-<li><a href="rest-api/detector-stop-.html">Stop genius framing and detector</a> : <code>PUT /detector/stop</code></li>
+<li><a href="rest-api/detector-start.html">Start autozoom (genius framing) and detector</a> : <code>PUT /detector/start</code></li>
+<li><a href="rest-api/detector-stop-.html">Stop autozoom (genius framing) and detector</a> : <code>PUT /detector/stop</code></li>
 <li><a href="rest-api/get-detections.html">Get detections</a> : <code>GET /detector/detections</code></li>
 </ul>
 <h2 id="upgrade">Upgrade</h2>
@@ -71,7 +71,7 @@
 <li><a href="rest-api/upgrade.html">Upgrade camera</a> : <code>POST /upgrade</code></li>
 <li><a href="rest-api/upgrade-status.html">Status of upgrade</a> : <code>GET /upgrade/status</code></li>
 </ul>
-<p>The server is using the huddly-sdk to communicate with the camer, and koa for the web application.</p>
+<p>The server is using the huddly-sdk to communicate with the camera, and koa for the web application.</p>
 
                    </div><div class="search-results">
     <div class="has-results">
