+++ docs/additional-documentation/rest-api/detector-stop.html
@@ -55,7 +55,7 @@
 <p><strong>Code</strong> : <code>200 OK</code></p>
 <p><strong>Content examples</strong></p>
 <div><pre class="line-numbers"><code class="language-json">{
-    &quot;status&quot;: &quot;Autozoom stopped!&quot;
+    &quot;status&quot;: &quot;Autozoom stopped and unsubscribed from detection events!&quot;
 }</code></pre></div>
                    </div><div class="search-results">
     <div class="has-results">
