+++ docs/additional-documentation/rest-api/get-detections.html
@@ -48,7 +48,7 @@
 
 
 <h1 id="get-detections">Get detections</h1>
-<p>Quires for the last detections that have come in.</p>
+<p>Queries for the last detections that have come in.</p>
 <p><strong>URL</strong> : <code>/detector/detections</code></p>
 <p><strong>Method</strong> : <code>GET</code></p>
 <h2 id="success-response">Success Response</h2>
