+++ docs/index.html
@@ -81,6 +81,7 @@ sdk.on(&#39;ATTACH&#39;, (d) =&gt; {
 <ul>
 <li><a href="http://developer.huddly.com/classes/HuddlySdk.html">SDK</a></li>
 <li><a href="http://developer.huddly.com/interfaces/IDeviceManager.html">IDeviceManager</a></li>
+<li><a href="http://developer.huddly.com/interfaces/IAutozoomCtl.html">Autozoom Controller</a></li>
 <li><a href="http://developer.huddly.com/interfaces/IDetector.html">Detector</a></li>
 </ul>
 <h2 id="repo">Repo</h2>
