+++ docs/interfaces/DetectorOpts.html
@@ -87,15 +87,15 @@
                             <ul class="index-list">
                                 <li>
                                             <span class="modifier">Optional</span>
+                                        <a href="#configDetectionsOnSubstream">configDetectionsOnSubstream</a>
+                                </li>
+                                <li>
+                                            <span class="modifier">Optional</span>
                                         <a href="#convertDetections">convertDetections</a>
                                 </li>
                                 <li>
                                             <span class="modifier">Optional</span>
                                         <a href="#objectFilter">objectFilter</a>
-                                </li>
-                                <li>
-                                            <span class="modifier">Optional</span>
-                                        <a href="#shouldAutoFrame">shouldAutoFrame</a>
                                 </li>
                             </ul>
                         </td>
@@ -108,6 +108,53 @@
 
             <section>
                 <h3 id="inputs">Properties</h3>
+                    <table class="table table-sm table-bordered">
+                        <tbody>
+                                <tr>
+                                    <td class="col-md-4">
+                                        <a name="configDetectionsOnSubstream"></a>
+                                        <span class="name"><b>configDetectionsOnSubstream</b><a href="#configDetectionsOnSubstream"><span class="icon ion-ios-link"></span></a></span>
+                                    </td>
+                                </tr>
+                                <tr>
+                                    <td class="col-md-4">
+                                        <code>configDetectionsOnSubstream:     <code><a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/boolean" target="_blank" >boolean</a></code>
+</code>
+                                    </td>
+                                </tr>
+
+
+                                    <tr>
+                                        <td class="col-md-4">
+                                            <i>Type : </i>    <code><a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/boolean" target="_blank" >boolean</a></code>
+
+                                        </td>
+                                    </tr>
+
+                                    <tr>
+                                        <td class="col-md-4">
+                                            <i>Optional</i>
+                                        </td>
+                                    </tr>
+
+
+
+
+                            <tr>
+                                <td class="col-md-4">
+                                    <div class="io-description"><p>This option sets up the <code>Detector</code> class to configure the
+camera in such a way that you can get detection data without
+the need of streaming the main stream using a third party streaming
+application on the host machine.</p>
+<p><strong>NOTE</strong> Always call <code>#destroy</code> method on <code>Detector</code> class when
+tearing down the application. the <code>#destroy</code> function will make
+sure that the camera stops the internal stream that emits detection
+data to the host.</p>
+</div>
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
                     <table class="table table-sm table-bordered">
                         <tbody>
                                 <tr>
@@ -193,47 +240,6 @@ Provide an empty list to retrieve all</p>
                             </tr>
                         </tbody>
                     </table>
-                    <table class="table table-sm table-bordered">
-                        <tbody>
-                                <tr>
-                                    <td class="col-md-4">
-                                        <a name="shouldAutoFrame"></a>
-                                        <span class="name"><b>shouldAutoFrame</b><a href="#shouldAutoFrame"><span class="icon ion-ios-link"></span></a></span>
-                                    </td>
-                                </tr>
-                                <tr>
-                                    <td class="col-md-4">
-                                        <code>shouldAutoFrame:     <code><a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/boolean" target="_blank" >boolean</a></code>
-</code>
-                                    </td>
-                                </tr>
-
-
-                                    <tr>
-                                        <td class="col-md-4">
-                                            <i>Type : </i>    <code><a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/boolean" target="_blank" >boolean</a></code>
-
-                                        </td>
-                                    </tr>
-
-                                    <tr>
-                                        <td class="col-md-4">
-                                            <i>Optional</i>
-                                        </td>
-                                    </tr>
-
-
-
-
-                            <tr>
-                                <td class="col-md-4">
-                                    <div class="io-description"><p>Set to false in case you want to get detections without
-enabling genius framing</p>
-</div>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
             </section>
     </div>
 
@@ -262,15 +268,6 @@ export default interface DetectorOpts {
   convertDetections?: DetectionConvertion;
 
   /**
-   * Set to false in case you want to get detections without
-   * enabling genius framing
-   *
-   * @type {Boolean}
-   * @memberof DetectorOpts
-   */
-  shouldAutoFrame?: boolean;
-
-  /**
    * Specify a list of labels for objects to detect, ie [&#x27;head&#x27;, &#x27;person&#x27;]
    * Provide an empty list to retrieve all
    *
@@ -278,6 +275,22 @@ export default interface DetectorOpts {
    * @memberof DetectorOpts
    */
   objectFilter?: Array&lt;String&gt;;
+
+  /**
+   * This option sets up the &#x60;Detector&#x60; class to configure the
+   * camera in such a way that you can get detection data without
+   * the need of streaming the main stream using a third party streaming
+   * application on the host machine.
+   *
+   * **NOTE** Always call &#x60;#destroy&#x60; method on &#x60;Detector&#x60; class when
+   * tearing down the application. the &#x60;#destroy&#x60; function will make
+   * sure that the camera stops the internal stream that emits detection
+   * data to the host.
+   *
+   * @type {boolean}
+   * @memberof DetectorOpts
+   */
+  configDetectionsOnSubstream?: boolean;
 }
 
 export { DetectionConvertion };
