+++ docs/overview.html
@@ -49,7 +49,7 @@
             <div class="card text-center">
                 <div class="card-block">
                     <h4 class="card-title"><span class="icon ion-md-information-circle-outline"></span></h4>
-                    <p class="card-text">7 Interfaces</p>
+                    <p class="card-text">9 Interfaces</p>
                 </div>
             </div>
         </div>
