+++ examples/azure_iot/demo_steps/detect.js
@@ -5,11 +5,33 @@ const DeviceApiUsb = require('@huddly/device-api-usb').default;
 const huddlySdk = new HuddlySdk(new DeviceApiUsb());
 
 huddlySdk.on(CameraEvents.ATTACH, async (deviceManager) => {
-  const detector = await deviceManager.getDetector();
-  detector.on(CameraEvents.DETECTIONS, (detections) => {
+  this.deviceManager = deviceManager;
+  this.autozoomCtl = this.deviceManager.getAutozoomCtl();
+  await this.autozoomCtl.init();
+  await this.autozoomCtl.start();
+  this.detector = this.deviceManager.getDetector();
+  this.detector.on(CameraEvents.DETECTIONS, (detections) => {
     console.log(detections);
   });
-  detector.start();
+  this.detector.init();
+});
+
+process.on('SIGINT', async () => {
+  // Keyboard interrupt signal
+
+  if (this.detector) {
+    await this.detector.destroy();
+  }
+
+  if(this.autozoomCtl) {
+    await this.autozoomCtl.stop();
+  }
+
+  if (this.deviceManager) {
+    await this.deviceManager.closeConnection();
+  }
+  console.log("\nApplication successfully closed!");
+  process.exit();
 });
 
 huddlySdk.init();
