+++ examples/rest/README.md
@@ -32,12 +32,12 @@ Here are the actions you can perform:
 * [Get camera info](rest-api/camera-info.html) : `GET /info`
 
 ## Detector
-* [Start genius framing and detector](rest-api/detector-start.html) : `PUT /detector/start`
-* [Stop genius framing and detector](rest-api/detector-stop-.html) : `PUT /detector/stop`
+* [Start autozoom (genius framing) and detector](rest-api/detector-start.html) : `PUT /detector/start`
+* [Stop autozoom (genius framing) and detector](rest-api/detector-stop-.html) : `PUT /detector/stop`
 * [Get detections](rest-api/get-detections.html) : `GET /detector/detections`
 
 ## Upgrade
 * [Upgrade camera](rest-api/upgrade.html) : `POST /upgrade`
 * [Status of upgrade](rest-api/upgrade-status.html) : `GET /upgrade/status`
 
-The server is using the huddly-sdk to communicate with the camer, and koa for the web application.
+The server is using the huddly-sdk to communicate with the camera, and koa for the web application.
