+++ src/components/device/boxfish.ts
@@ -15,6 +15,9 @@ import { EventEmitter } from 'events';
 import { createBoxfishUpgrader } from './../upgrader/boxfishUpgraderFactory';
 import BoxfishUpgrader from './../upgrader/boxfishUpgrader';
 import InterpolationParams from './../../interfaces/InterpolationParams';
+import AutozoomCtlOpts from './../../interfaces/IAutozoomCtlOpts';
+import IAutozoomCtl from './../../interfaces/IAutozoomCtl';
+import AutozoomCtl from './../autozoomCtl';
 
 const MAX_UPGRADE_ATTEMPT = 3;
 
@@ -235,7 +238,11 @@ export default class Boxfish extends UvcBaseDevice implements IDeviceManager {
       }
   }
 
-   getDetector(opts?: DetectorOpts): IDetector {
+  getAutozoomCtl(opts: AutozoomCtlOpts): IAutozoomCtl {
+    return new AutozoomCtl(this, this.logger, opts);
+  }
+
+  getDetector(opts?: DetectorOpts): IDetector {
     return new Detector(this, this.logger, opts);
   }
