+++ src/interfaces/IDetector.md
@@ -1,20 +1,29 @@
 ## How to use the detector interface!
 
-Below you can find an example on how to get started with Genius Framing on IQ. Follow the steps below for configuring Genius Framing on the camera
-and subscribing for detection and framing information!
+In order to use the detector interface you need to make sure that the autozoom feature is enabled on the camera (see the [IAutozoomCtl](http://developer.huddly.com/interfaces/IAutozoomCtl.html) interface for documentation on how to do it).
+
+The detector interface allows you to subscribe to detection and/or framing data from the camera. It is also possible to configure the detector in a way where detection data can be retrieved even without streaming on host using a streaming application. Examples below show how the detector can be used.
+
 ```javascript
 /* 
-We suppose that the SDK has been initialized previously
-and the camera is attached and an instance of IDeviceManager
-is created!
+We suppose that the sdk has been initialized and autozoom feature is enabled
+and running.
 */
-const detector = await cameraManager.getDetector();
+...
+
+/**
+ * Get an instance of detector implementation from the
+ * camera manager.
+ */
+const detector = cameraManager.getDetector();
+
+// Always call `init`
 await detector.init();
-await detector.start();
 
 /*
-In order to subscribe for detection and framing information
-proceed with the following code snippet
+Detector class implements EventEmitter and therefor implements
+all the event emitter methods. Below we show how to listen
+to detector events such as DETECTIONS and FRAMING.
 */
 
 detector.on('DETECTIONS', (detectionList) => {
@@ -26,37 +35,45 @@ detector.on('FRAMING', (framingInfo) => {
 });
 
 
-/* To stop Genius Framing on IQ, use the #stop method */
-await detector.stop();
+/**
+ * When you are done call `destroy` method to properly teardown
+ * the detection instance.
+ */
+await detector.destroy();
 
+...
 ```
 
 ## Detection data without streaming
-Wondering whether it is possible getting detection (but not framing) data from the camera without streaming on a third party streaming application? The answer is yes! This new feature has been release with version 0.3.2 of the SDK and the follow code snippet illustrates the usage of the new feature:
+Wondering how to configure detector to get detection data without streaming main? The new feature has been added on release 0.4.0 and can be set up as the following:
 
 ``` javascript
-/* 
-We suppose that the SDK has been initialized previously
-and the camera is attached and an instance of IDeviceManager
-is created!
-*/
 ...
 
-const detector = await cameraManager.getDetector();
-await detector.init();
+/**
+ * The `getDetector` method accepts options documented on
+ * `IDetectorOpts` interface. One of the options allows you
+ * to configure the detector to get detection data without
+ * having to stream main stream on the host machine.
+ */
+const detector = await cameraManager.getDetector({
+  configDetectionsOnSubstream: true,
+});
 
-await detector.detectorStart(); // Call #detectorStart instead of #start for getting detection data without streaming on host
+// Call init
+await detector.init();
 
 // Setup detection event listener
 detector.on('DETECTIONS', (detectionList) => {
   // Your code here
 });
 
-// !!! IMPORTANT !!! Make sure you call #detectorStop when you are done with detection data
-await detector.detectorStop();
+// !!! IMPORTANT !!! Make sure you call `destroy` when you are done with detection data
+await detector.destroy();
 ...
 ```
-**!! NOTE !!** When tearing down your application which is using the detector object, it is important that you call `detectorStop` method before closing the host application. Not doing so will leave the camera streaming after you have closed the application which might lead to confusion. The camera will stop generating detection data at some point when it detects that no host application is consuming its data. However, it is strongly recommended to properly teardown the detector instance from the host application.
+**!! NOTE !!** When tearing down your application which is using the detector object, it is important that you call `destroy` method before closing the host application. Not doing so will leave the camera streaming internally after you have closed the application which might lead to confusion. The camera will stop generating detection data at some point when it detects that no host application is consuming its data. However, it is strongly recommended to properly teardown the detector instance from the host application.
+
 ## Detector Events
 The detector class can emit the following events:
