+++ src/interfaces/IDetectorOpts.ts
@@ -21,15 +21,6 @@ export default interface DetectorOpts {
   convertDetections?: DetectionConvertion;
 
   /**
-   * Set to false in case you want to get detections without
-   * enabling genius framing
-   *
-   * @type {Boolean}
-   * @memberof DetectorOpts
-   */
-  shouldAutoFrame?: boolean;
-
-  /**
    * Specify a list of labels for objects to detect, ie ['head', 'person']
    * Provide an empty list to retrieve all
    *
@@ -37,6 +28,22 @@ export default interface DetectorOpts {
    * @memberof DetectorOpts
    */
   objectFilter?: Array<String>;
+
+  /**
+   * This option sets up the `Detector` class to configure the
+   * camera in such a way that you can get detection data without
+   * the need of streaming the main stream using a third party streaming
+   * application on the host machine.
+   *
+   * **NOTE** Always call `#destroy` method on `Detector` class when
+   * tearing down the application. the `#destroy` function will make
+   * sure that the camera stops the internal stream that emits detection
+   * data to the host.
+   *
+   * @type {boolean}
+   * @memberof DetectorOpts
+   */
+  configDetectionsOnSubstream?: boolean;
 }
 
 export { DetectionConvertion };
